feat: refactor toast API to use core package and enhance send functionality
