--- surf-api-bukkit/surf-api-bukkit-api/src/main/kotlin/dev/slne/surf/surfapi/bukkit/api/toast/Toast.kt
@@ -1,13 +1,0 @@
-package dev.slne.surf.surfapi.bukkit.api.toast
-
-import net.kyori.adventure.text.Component
-import org.bukkit.Material
-import org.bukkit.entity.Player
-
-interface Toast {
-    val icon: Material
-    val text: Component
-    val style: ToastStyle
-
-    fun send(player: Player)
-}
