--- surf-api-bukkit/surf-api-bukkit-api/src/main/kotlin/dev/slne/surf/surfapi/bukkit/api/toast/ToastStyle.kt
@@ -1,7 +1,0 @@
-package dev.slne.surf.surfapi.bukkit.api.toast
-
-enum class ToastStyle {
-    TASK,
-    CHALLENGE,
-    GOAL
-}
