+++ surf-api-bukkit/surf-api-bukkit-api/src/main/kotlin/dev/slne/surf/surfapi/bukkit/api/util/bukkit-util.kt
@@ -7,8 +7,8 @@ import com.github.shynixn.mccoroutine.folia.entityDispatcher
 import com.github.shynixn.mccoroutine.folia.regionDispatcher
 import dev.slne.surf.surfapi.bukkit.api.SurfBukkitApi
 import dev.slne.surf.surfapi.bukkit.api.surfBukkitApi
-import dev.slne.surf.surfapi.bukkit.api.toast.Toast
-import dev.slne.surf.surfapi.bukkit.api.toast.ToastBuilder
+import dev.slne.surf.surfapi.core.api.toast.Toast
+import dev.slne.surf.surfapi.core.api.toast.ToastBuilder
 import dev.slne.surf.surfapi.core.api.util.getCallerClass
 import dev.slne.surf.surfapi.core.api.util.mutableLong2ObjectMapOf
 import dev.slne.surf.surfapi.core.api.util.mutableObjectListOf
@@ -274,5 +274,8 @@ suspend fun World.getBlockAtAsync(pos: BlockPosition): Block {
     }
 }
 
-fun Player.sendToast(toast: Toast) = toast.send(this)
-fun Player.sendToast(block: ToastBuilder.() -> Unit) = surfBukkitApi.createToast(block).send(this)
+fun Player.sendToast(toast: Toast) = toast.send(this.uniqueId)
+fun Player.sendToast(block: ToastBuilder.() -> Unit) =
+    surfBukkitApi.createToast(block).send(this.uniqueId)
+
+fun Toast.send(player: Player) = send(player.uniqueId)
