+++ surf-api-bukkit/surf-api-bukkit-plugin-test/src/main/kotlin/dev/slne/surf/surfapi/bukkit/test/command/subcommands/ToastTest.kt
@@ -1,19 +1,21 @@
 package dev.slne.surf.surfapi.bukkit.test.command.subcommands
 
+import com.github.retrooper.packetevents.protocol.item.type.ItemTypes
 import dev.jorel.commandapi.CommandAPICommand
 import dev.jorel.commandapi.kotlindsl.playerExecutor
 import dev.slne.surf.surfapi.bukkit.api.surfBukkitApi
-import dev.slne.surf.surfapi.bukkit.api.toast.ToastStyle
-import dev.slne.surf.surfapi.bukkit.api.toast.toast
+import dev.slne.surf.surfapi.bukkit.api.util.send
 import dev.slne.surf.surfapi.bukkit.api.util.sendToast
 import dev.slne.surf.surfapi.core.api.messages.adventure.buildText
+import dev.slne.surf.surfapi.core.api.toast.ToastStyle
+import dev.slne.surf.surfapi.core.api.toast.toast
 import org.bukkit.Material
 
 class ToastTest(name: String) : CommandAPICommand(name) {
     init {
         playerExecutor { player, _ ->
             player.sendToast {
-                icon(Material.DIAMOND)
+                icon(ItemTypes.DIAMOND)
 
                 text {
                     info("Dsl-Extension Test")
@@ -23,7 +25,7 @@ class ToastTest(name: String) : CommandAPICommand(name) {
             }
 
             player.sendToast(toast {
-                icon(Material.GOLD_INGOT)
+                icon(ItemTypes.GOLD_INGOT)
 
                 text {
                     info("Toast-Builder-Dsl Test")
