+++ surf-api-core/surf-api-core-api/src/main/kotlin/dev/slne/surf/surfapi/core/api/toast/ToastStyle.kt
@@ -1,0 +1,7 @@
+package dev.slne.surf.surfapi.core.api.toast
+
+enum class ToastStyle {
+    TASK,
+    CHALLENGE,
+    GOAL
+}
